traits: refactor traits for HID class-specific behavior
Changes to using traits for HID class-specific behavior, instead of
structs that implement a single generic `KeyboardOps` trait. This allows
for a single concrete implementation for storing all state information,
while still being able to call class-specific functionality.
